+++ src/pages/ProfilePage/index.js
@@ -1,6 +1,9 @@
 
-import {Box, Container, Grid} from '@mui/material/';
+import {Box, Container, Grid, Typography,} from '@mui/material/';
 import ProfileCard from './profileCard';
+import Experience from './List';
+import FriendsList from './Friends';
+
 
 
 
@@ -11,9 +14,13 @@ export default function ProfilePage() {
   container
   direction="column"
   justifyContent="space-evenly"
-  alignItems="flex-start"
+  alignItems="center"
+  sx={{mt: 3, mb: 3}}
 >
   <ProfileCard />
+  <Experience />
+  <FriendsList/>
+  
 </Grid>
 </>
   );
