+++ src/pages/ProfilePage/profileCard.js
@@ -3,31 +3,16 @@ import Card from '@mui/material/Card';
 import CardHeader from '@mui/material/CardHeader';
 import CardContent from '@mui/material/CardContent';
 import CardActions from '@mui/material/CardActions';
-import Collapse from '@mui/material/Collapse';
 import Avatar from '@mui/material/Avatar';
 import IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
 import { red } from '@mui/material/colors';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+
 import MoreVertIcon from '@mui/icons-material/MoreVert';
 import { GitHub, Email, LinkedIn } from '@mui/icons-material';
 import * as React from 'react';
 import { styled } from '@mui/material/styles';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemAvatar from '@mui/material/ListItemAvatar';
-import ListItemText from '@mui/material/ListItemText';
-import DeleteIcon from '@mui/icons-material/Delete';
-import Experience from './List';
 
-
-function generate(element) {
-  return [0, 1, 2].map((value) =>
-    React.cloneElement(element, {
-      key: value,
-    }),
-  );
-}
 
 
 
@@ -50,7 +35,7 @@ export default function ProfileCard() {
   };
 
   return (
-    <Card sx={{ maxWidth: 752 }}  >
+    <Card sx={{ maxWidth: 500 }}  >
       <CardHeader
         avatar={
           <Avatar sx={{ bgcolor: red[500] }} aria-label="recipe">
@@ -65,7 +50,7 @@ export default function ProfileCard() {
         title="Brian M"
         subheader="Full-Stack Developer"
       />
-    
+  
       <CardContent>
         <Typography variant="subtitle1">
         About Me:
@@ -90,43 +75,7 @@ export default function ProfileCard() {
         <IconButton aria-label="LinkedIn">
           <LinkedIn />
         </IconButton>
-        <ExpandMore
-          expand={expanded}
-          onClick={handleExpandClick}
-          aria-expanded={expanded}
-          aria-label="show more"
-        >
-          <ExpandMoreIcon />
-        </ExpandMore>
       </CardActions>
-      <Collapse in={expanded} timeout="auto" unmountOnExit>
-        <CardContent>
-            <Experience />
-        </CardContent>
-      </Collapse>
-      <Typography sx={{ mt: 1, mb: 2 }} variant="h6" component="div">
-            Friends List
-          </Typography>
-            <List>
-              {generate(
-                <ListItem
-                  secondaryAction={
-                    <IconButton edge="end" aria-label="delete">
-                      <DeleteIcon />
-                    </IconButton>
-                  }
-                >
-                  <ListItemAvatar>
-                    <Avatar>
-                      <Avatar />
-                    </Avatar>
-                  </ListItemAvatar>
-                  <ListItemText
-                    primary="Friends"
-                  />
-                </ListItem>,
-              )}
-            </List>
     </Card>
   );
 }
